requirements.txt
# Landscaping Design Bot - Dependencies

# Chat Interface
chainlit>=1.0.0

# Google Gemini AI
google-generativeai>=0.8.0

# HTTP Client
httpx>=0.27.0

# Image Processing
Pillow>=10.0.0

# Environment Variables
python-dotenv>=1.0.0
